aux window - drop buggy `AuxWindowUnsupported` capability
This needs a different approach, likely on the level of custom editors itself. Given custom editors are to implement backups, moving between windows should actually not result in data loss in most cases.

For custom editors that ignore backups, a better approach would be to throw an error or force to save right before a `webview` is about to get disposed because it moved to another window.